requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.0
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2